+++ packages/template-retail-react-app/app/pages/login/index.jsx
@@ -23,6 +23,7 @@ import {useForm} from 'react-hook-form'
 import {useRouteMatch} from 'react-router'
 import {useLocation} from 'react-router-dom'
 import useEinstein from '@salesforce/retail-react-app/app/hooks/use-einstein'
+import useDatacloud from '@salesforce/retail-react-app/app/hooks/use-datacloud'
 import LoginForm from '@salesforce/retail-react-app/app/components/login'
 import PasswordlessEmailConfirmation from '@salesforce/retail-react-app/app/components/email-confirmation/index'
 import {
@@ -56,6 +57,7 @@ const Login = ({initialView = LOGIN_VIEW}) => {
     const queryParams = new URLSearchParams(location.search)
     const {path} = useRouteMatch()
     const einstein = useEinstein()
+    const datacloud = useDatacloud()
     const {isRegistered, customerType} = useCustomerType()
     const login = useAuthHelper(AuthHelpers.LoginRegisteredUserB2C)
     const loginPasswordless = useAuthHelper(AuthHelpers.LoginPasswordlessUser)
@@ -184,6 +186,7 @@ const Login = ({initialView = LOGIN_VIEW}) => {
     /**************** Einstein ****************/
     useEffect(() => {
         einstein.sendViewPage(location.pathname)
+        datacloud.sendViewPage(location.pathname)
     }, [])
 
     return (