+++ packages/template-retail-react-app/config/mocks/default.js
@@ -95,6 +95,10 @@ module.exports = {
             // This is temporary and is meant as a placeholder until there is a mechanism for reading
             // the is_production property from an MRT target
             isProduction: false
+        },
+        dataCloudAPI: {
+            appSourceId: '23df7335-2e9d-4fbc-bc34-7e93649e69b7',
+            tenantId: '5zqheixqu9vji7spdkzxwh4hpz'
         }
     },
     // This list contains server-side only libraries that you don't want to be compiled by webpack