+++ packages/template-retail-react-app/CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### Accessibility Improvements
 - [a11y] Fix LinkList component to follow a11y practise [#2098])(https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2098)
+- Test Node 22
 
 ## v5.0.0 (Oct 28, 2024)